kin_news_core/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
DAWG-Python==0.7.2
ddt==1.4.2
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
docopt==0.6.2
drf-yasg==1.21.4
emoji==2.1.0
flake8==5.0.4
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
nltk==3.7
numpy==1.23.3
oauthlib==3.1.0
packaging==21.3
Pillow==8.2.0
psycopg2-binary==2.8.6
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.20
pyflakes==2.5.0
PyJWT==2.0.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2021.1
regex==2022.9.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.1
six==1.15.0
Telethon==1.25.2
threadpoolctl==3.1.0
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.4
